Moved EEPROM bytes to end of HomeKit-ESP8266 EEPROM section
The HomeKit-ESP8266 library requires the first 1408 bytes of EEPROM to be used by the library.  Every time this program would launch, the HomeKit configuration bytes were being overwritten, thus triggering the library to create a new HomeKit configuration.  This was resolved by moving this project's EEPROM bytes to the end of the library's EEPROM section.