Fixed bug #10
Apply delay after showing color data on NeoPixels.  After referencing example projects, I noticed a delay was being applied after color data was pushed to the NeoPixels (20-50ms).  Originally, this was not applied to make the fading transition between NeoPixels as smooth as possible, but this has seemed to cause some problems while connected to Wi-Fi (probably interrupt-related).  The code was updated to make sure it is executed once every 50ms.